+++ .github/TEST_FAILURE_REPORT_TEMPLATE.md
@@ -5,6 +5,6 @@ labels: ['type::bug', 'type::testing', 'source::auto']
 
 The {{ workflow }} workflow failed on {{ date | date("YYYY-MM-DD HH:mm") }} UTC
 
-Full run: https://github.com/anaconda/anaconda-conda-tos/actions/runs/{{ env.RUN_ID }}
+Full run: https://github.com/anaconda/conda-anaconda-tos/actions/runs/{{ env.RUN_ID }}
 
 (This post will be updated if another test fails today, as long as this issue remains open.)
