+++ pyproject.toml
@@ -25,11 +25,11 @@ dependencies = [
 description = "Conda subcommand to view, accept, and interact with a channel's Terms of Service (ToS)."
 dynamic = ["version"]
 license = {file = "LICENSE"}
-name = "anaconda-conda-tos"
+name = "conda-anaconda-tos"
 requires-python = ">=3.9"
 
 [project.entry-points.conda]
-anaconda-conda-tos = "anaconda_conda_tos.plugin"
+conda-anaconda-tos = "conda_anaconda_tos.plugin"
 
 [tool.coverage.report]
 exclude_lines = [
@@ -47,8 +47,11 @@ sort = "Miss"
 # store relative paths in coverage information
 relative_files = true
 
+[tool.hatch.build]
+include = ["conda_anaconda_tos"]
+
 [tool.hatch.build.hooks.vcs]
-version-file = "anaconda_conda_tos/_version.py"
+version-file = "conda_anaconda_tos/_version.py"
 
 [tool.hatch.version]
 source = "vcs"
@@ -59,7 +62,7 @@ local_scheme = "dirty-tag"
 [tool.pytest.ini_options]
 addopts = [
   "--color=yes",
-  # "--cov=anaconda_conda_tos",  # passed in test runner scripts instead (avoid debugger)
+  # "--cov=conda_anaconda_tos",  # passed in test runner scripts instead (avoid debugger)
   "--cov-report=term",  # print summary table to screen
   "--cov-report=xml",  # for codecov/codecov-action upload
   "--tb=native",
@@ -104,6 +107,3 @@ select = [
   "W",  # pycodestyle warnings
   "YTT",  # flake8-2020
 ]
-
-[tool.setuptools]
-packages = ["anaconda_conda_tos"]
