+++ recipe/meta.yaml
@@ -1,4 +1,4 @@
-{% set name = "anaconda-conda-tos" %}
+{% set name = "conda-anaconda-tos" %}
 
 package:
   name: {{ name|lower }}
@@ -31,7 +31,7 @@ test:
     - pip
     - conda >24.9.2
   imports:
-    - anaconda_conda_tos
+    - conda_anaconda_tos
   commands:
     - pip check
     - conda tos --version
