+++ tests/test_local.py
@@ -12,15 +12,15 @@ from conda.common.compat import on_win
 from conda.models.channel import Channel
 from pydantic import ValidationError
 
-from anaconda_conda_tos.exceptions import CondaToSMissingError, CondaToSPermissionError
-from anaconda_conda_tos.local import (
+from conda_anaconda_tos.exceptions import CondaToSMissingError, CondaToSPermissionError
+from conda_anaconda_tos.local import (
     get_local_metadata,
     get_local_metadatas,
     read_metadata,
     write_metadata,
 )
-from anaconda_conda_tos.models import LocalPair, LocalToSMetadata, RemoteToSMetadata
-from anaconda_conda_tos.path import get_metadata_path
+from conda_anaconda_tos.models import LocalPair, LocalToSMetadata, RemoteToSMetadata
+from conda_anaconda_tos.path import get_metadata_path
 
 if TYPE_CHECKING:
     from pathlib import Path
